Separate out HelpRequestCallFactory methods
